requirements.txt
asgiref==3.10.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==44.0.3
dj-rest-auth==7.0.1
Django==5.2.7
django-allauth==65.12.1
django-mfa2==3.1.1
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
ecdsa==0.19.1
fido2==1.2.0
idna==3.11
pyasn1==0.6.1
pycparser==2.23
PyJWT==2.10.1
pyotp==2.9.0
python-decouple==3.8
python-dotenv==1.2.1
python-jose==3.5.0
python-u2flib-server==5.0.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.15.0
ua-parser==1.0.1
ua-parser-builtins==0.18.0.post1
urllib3==2.5.0
user-agents==2.2.0
